+++ lax/interfaces/patchinterface.cc
@@ -566,7 +566,7 @@ void PatchData::dump_out(FILE *f,int indent,int what,Laxkit::DumpContext *contex
 	}
 
 	fprintf(f,"%smatrix %.10g %.10g %.10g %.10g %.10g %.10g\n",
-			spc,m(0),m(1),m(2),m(3),m(4),m(5));
+			spc,m((int)0),m(1),m(2),m(3),m(4),m(5));
 	fprintf(f,"%sgriddivisions %d\n",spc,griddivisions);
 	if (style&PATCH_SMOOTH) fprintf(f,"%sstyle smooth\n",spc);
 	
@@ -621,7 +621,7 @@ Laxkit::Attribute *PatchData::dump_out_atts(Laxkit::Attribute *att,int what,Laxk
 	}
 
 	att->pushStr("matrix",-1, "%.10g %.10g %.10g %.10g %.10g %.10g",
-			m(0),m(1),m(2),m(3),m(4),m(5));
+			m((int)0),m(1),m(2),m(3),m(4),m(5));
 	att->push("griddivisions",griddivisions);
 	if (style&PATCH_SMOOTH) att->push("style", "smooth");
